+++ src/js/casc/casc-source.js
@@ -117,7 +117,7 @@ class CASC {
 		core.view.$watch('config.listfileShowFileDataIDs', () => {
 			const showFileIDs = core.view.config.listfileShowFileDataIDs;
 			core.view.listfileTextures = listfile.getFilenamesByExtension('.blp', showFileIDs);
-			core.view.listfileSounds = listfile.getFilenamesByExtension(['.ogg', '.mp3'], showFileIDs);
+			core.view.listfileSounds = listfile.getFilenamesByExtension(['.ogg', '.mp3', '.unk_sound'], showFileIDs);
 			core.view.listfileVideos = listfile.getFilenamesByExtension('.avi', showFileIDs);
 			core.view.listfileText = listfile.getFilenamesByExtension(['.txt', '.lua', '.xml', '.sbt', '.wtf', '.htm', '.toc', '.xsd'], showFileIDs);
 			core.view.listfileModels = listfile.getFilenamesByExtension(modelExt, showFileIDs);